# Merge a lot of optimize from bytes.Buffer